requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.4
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-admin-rangefilter==0.11.2
django-admin-sortable2==2.1.10
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==3.3.1
django-pandas==0.6.6
django-timezone-field==6.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
frozenlist==1.4.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.16.1
jedi==0.19.1
kombu==5.3.2
llvmlite==0.41.1
MarkupPy==1.14
matplotlib-inline==0.1.6
msgpack==1.0.7
multidict==6.0.4
numba==0.58.1
numpy==1.26.0
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.3.0
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pyotp==2.9.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
selenium==4.14.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.3
TA-Lib==0.4.28
tablib==3.5.0
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.11.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
zope.interface==6.1
